setup GH actions, clean up samples, add missing provides, minor tweaks in poms
