+++ .gitignore
@@ -1,3 +1,4 @@
 target/
 */target/
 .m2/
+/.idea/
